Added tailwind to match modern SaaS products.
