+++ frontend/package.json
@@ -24,6 +24,7 @@
     "@fastify/middie": "^9.0.3",
     "@fastify/static": "^8.1.1",
     "@quasar/extras": "^1.16.4",
+    "@tailwindcss/postcss": "^4.0.9",
     "axios": "^1.2.1",
     "express": "^4.21.2",
     "fastify": "^5.2.1",
@@ -44,11 +45,13 @@
     "@vue/eslint-config-prettier": "^10.1.0",
     "@vue/eslint-config-typescript": "^14.1.3",
     "@vue/test-utils": "^2.4.4",
-    "autoprefixer": "^10.4.2",
+    "autoprefixer": "^10.4.20",
     "eslint": "^9.14.0",
     "eslint-plugin-vue": "^9.30.0",
     "globals": "^15.12.0",
+    "postcss": "^8.5.3",
     "prettier": "^3.3.3",
+    "tailwindcss": "^4.0.9",
     "typescript": "~5.5.3",
     "vite-plugin-checker": "^0.8.0",
     "vitest": "^3.0.0",
