+++ frontend/postcss.config.js
@@ -1,7 +1,7 @@
 // https://github.com/michael-ciniawsky/postcss-load-config
 
-import autoprefixer from 'autoprefixer'
-// import rtlcss from 'postcss-rtlcss'
+import autoprefixer from 'autoprefixer';
+import tailwindcss from '@tailwindcss/postcss';
 
 export default {
   plugins: [
@@ -15,15 +15,10 @@ export default {
         'last 4 Android versions',
         'last 4 ChromeAndroid versions',
         'last 4 FirefoxAndroid versions',
-        'last 4 iOS versions'
-      ]
+        'last 4 iOS versions',
+      ],
     }),
-
-    // https://github.com/elchininet/postcss-rtlcss
-    // If you want to support RTL css, then
-    // 1. yarn/pnpm/bun/npm install postcss-rtlcss
-    // 2. optionally set quasar.config.js > framework > lang to an RTL language
-    // 3. uncomment the following line (and its import statement above):
-    // rtlcss()
-  ]
-}
+    // https://github.com/tailwindlabs/tailwindcss
+    tailwindcss({}),
+  ],
+};
