+++ frontend/src/components/Misc/ThemeSwitcher.vue
@@ -15,7 +15,7 @@
 <script lang="ts" setup>
 import { computed } from 'vue';
 import { Dark } from 'quasar';
-import { toggleTheme } from 'src/composables/useTheme';
+import { toggleTheme } from 'src/utils/themeUtil';
 
 const isDark = computed(() => Dark.isActive);
